+++ src/main/java/com/nhnacademy/front/shop/couponStore/dto/CouponStatus.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.front.shop.couponStore.client.dto;
+package com.nhnacademy.front.shop.couponStore.dto;
 
 public enum CouponStatus {
     READY,
