+++ src/main/java/com/nhnacademy/front/shop/couponStore/dto/CouponStoreDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.front.shop.couponStore.client.dto;
+package com.nhnacademy.front.shop.couponStore.dto;
 
 import java.time.LocalDateTime;
