+++ src/main/java/com/nhnacademy/front/shop/couponStore/dto/DiscountType.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.front.shop.couponStore.client.dto;
+package com.nhnacademy.front.shop.couponStore.dto;
 
 public enum DiscountType {
     RATE,    // 정률 할인 (예: 10%)
